+++ protected/controllers/AdminExternalLinkController.php
@@ -51,19 +51,27 @@ class AdminExternalLinkController extends Controller
 	 */
 	public function actionCreate()
 	{
-		$model=new ExternalLink;
+		$model = new ExternalLink;
 
 		// Uncomment the following line if AJAX validation is needed
 		// $this->performAjaxValidation($model);
 
-		if(isset($_POST['ExternalLink']))
+		if ($attr = Yii::$app->request->post('ExternalLink'))
 		{
-			$model->attributes=$_POST['ExternalLink'];
-			if($model->save())
-				$this->redirect(array('view','id'=>$model->id));
+			$model->attributes = $attr;
+            $model->scenario='create';
+
+            $type = ExternalLinkType::model()->findByPk($attr['external_link_type_id']);
+            if ($type && $type->can_self_referred) {
+                $this->setRelatedId($model);
+            }
+
+            if (!$model->getErrors() && $model->save()) {
+                return $this->redirect(array('view','id'=>$model->id));
+            }
 		}
 
-		$this->render('create',array(
+		return $this->render('create',array(
 			'model'=>$model,
 		));
 	}
@@ -187,14 +195,34 @@ class AdminExternalLinkController extends Controller
 		// Uncomment the following line if AJAX validation is needed
 		// $this->performAjaxValidation($model);
 
-		if(isset($_POST['ExternalLink']))
+		if ($attr = Yii::$app->request->post('ExternalLink'))
 		{
-			$model->attributes=$_POST['ExternalLink'];
-			if($model->save())
-				$this->redirect(array('view','id'=>$model->id));
+			$model->attributes = $attr;
+            $model->scenario='update';
+            $relatedId = null;
+
+            $type = ExternalLinkType::model()->findByPk($attr['external_link_type_id']);
+            if ($type && $type->can_self_referred) {
+                $relatedId = $model->related_id;
+
+                $this->setRelatedId($model);
+            }
+
+            if (!$model->getErrors() && $model->save()) {
+                if ($relatedId) {
+                    $related = ExternalLink::model()->findByPk($relatedId);
+                    $count = $related->referencedBy;
+
+                    if ($related && !$count) {
+                        $related->is_referred = false;
+                        $related->save();
+                    }
+                }
+                return $this->redirect(array('view','id'=>$model->id));
+            }
 		}
 
-		$this->render('update',array(
+		return $this->render('update',array(
 			'model'=>$model,
 		));
 	}
@@ -307,4 +335,34 @@ class AdminExternalLinkController extends Controller
                  Util::returnJSON(array("success"=>false,"message"=>Yii::t("app", "Delete Error.")));
             }
         }
+
+    private function setRelatedId(ExternalLink $model): void
+    {
+        $softwareArchiveService = Yii::$app->get('softwareArchive');
+        $origin = $softwareArchiveService->getOriginUrl($model->url);
+
+        if (!$origin) {
+            $model->addError('error', 'Invalid archived link - no origin url found');
+
+            return;
+        }
+
+        $users = User::model()->findAll($criteria);
+        $externalLink = ExternalLink::model()->findByAttributes(
+            [
+                'url' => $origin,
+                'dataset_id' => $model->dataset_id,
+
+            ]
+        );
+
+        if ($externalLink) {
+            $externalLink->is_referred = true;
+            if ($externalLink->save()) {
+                //todo
+            }
+        }
+
+        $model->related_id = $externalLink ?  $externalLink->id : null;
+    }
 }